requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
black==23.11.0
celery==5.3.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
flake8==6.1.0
iniconfig==2.3.0
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==1.24.3
openpyxl==3.1.2
packaging==25.0
pandas==1.5.3
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
redis==5.0.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
dj-database-url==2.1.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
pytest-cov==4.1.0
